refactor(CardModal): move list of Comments to distinct CommentList component
