+++ src/components/CardModal/components/Comment/index.ts
@@ -1,1 +1,2 @@
 export { default as Comment } from "./Comment";
+export { default as CommentList} from "./CommentList";
