+++ src/components/CardModal/components/index.tsx
@@ -1,3 +1,4 @@
+export { CommentList } from "./Comment";
 export { Comment } from "./Comment";
 export { Description } from "./Description";
 export { NewComment } from "./NewComment";
